Finish Tutorial 28 - Putting everything Together

+++ app.js
@@ -1,6 +1,7 @@
 const express = require("express");
 const bodyParser = require("body-parser");
 const mongoose = require("mongoose");
+const _ = require("lodash");
 // const date = require(__dirname + "/date.js");
 
 const app = express();
@@ -10,7 +11,8 @@ app.set("view engine", "ejs");
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(express.static("public"));
 
-mongoose.connect("mongodb://192.168.56.11:27017/todolistDB");
+// mongoose.connect("mongodb://192.168.56.11:27017/todolistDB");
+mongoose.connect("mongodb://localhost:27017/todolistDB");
 
 const itemsSchema = new mongoose.Schema({
   name: {
@@ -71,8 +73,8 @@ app.get("/", function (req, res) {
 });
 
 app.get("/:customListName", (req, res) => {
-  const customListName = req.params.customListName;
-  console.log(customListName);
+  const customListName = _.capitalize(req.params.customListName);
+  //  console.log(customListName);
 
   List.findOne({ name: customListName }, (err, foundList) => {
     if (!err) {
@@ -84,24 +86,34 @@ app.get("/:customListName", (req, res) => {
         });
         list.save();
         res.redirect("/" + customListName);
+      } else {
+        res.render("list", {
+          listTitle: foundList.name,
+          newListItems: foundList.items,
+        });
       }
-    } else {
-      res.render("list", {
-        listTitle: foundList.name,
-        newListItems: foundList.items,
-      });
     }
   });
 });
 
 app.post("/", function (req, res) {
   const itemName = req.body.newItem;
+  const listName = req.body.list;
 
   const item = new Item({
     name: itemName,
   });
-  item.save();
-  res.redirect("/");
+
+  if (listName === "Today") {
+    item.save();
+    res.redirect("/");
+  } else {
+    List.findOne({ name: listName }, (err, foundList) => {
+      foundList.items.push(item);
+      foundList.save();
+      res.redirect("/" + listName);
+    });
+  }
 
   // if (req.body.list === "Work") {
   //   workItems.push(item);
@@ -113,30 +125,32 @@ app.post("/", function (req, res) {
 });
 
 app.post("/delete", (req, res) => {
-  checkedItemId = req.body.checkbox;
+  const checkedItemId = req.body.checkbox;
+  const listName = req.body.listName;
 
-  Item.findByIdAndRemove(checkedItemId, (err) => {
-    if (err) {
-      console.log(err);
-    } else {
-      res.redirect("/");
-      // console.log("Success Deleting Item.");
-    }
-  });
-});
-
-app.get("/work", function (req, res) {
-  res.render("list", { listTitle: "Work List", newListItems: workItems });
+  if (listName === "Today") {
+    Item.findByIdAndRemove(checkedItemId, (err) => {
+      if (err) {
+        console.log(err);
+      } else {
+        res.redirect("/");
+      }
+    });
+  } else {
+    List.findOneAndUpdate(
+      { name: listName },
+      { $pull: { items: { _id: checkedItemId } } },
+      (err, foundList) => {
+        if (!err) {
+          res.redirect("/" + listName);
+        }
+      }
+    );
+  }
 });
 
 app.get("/about", function (req, res) {
   res.render("about");
-});
-
-app.post("/work", function (req, res) {
-  const item = req.body.newItem;
-  workItems.push(item);
-  res.redirect("/work");
 });
 
 app.listen(3000, function () {
